Renamed action creator files to what it should be
