+++ package.json
@@ -141,7 +141,7 @@
     "concurrently": "^5.1.0",
     "copy-webpack-plugin": "^4.5.2",
     "cross-env": "^5.2.1",
-    "css-loader": "^0.28.11",
+    "css-loader": "^2.1.1",
     "electron": "^2.0.7",
     "electron-builder": "^20.44.4",
     "electron-devtools-installer": "^2.2.4",
