+++ src/components/Props.tsx
@@ -7,7 +7,7 @@ import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Select from '@material-ui/core/Select';
 import InputLabel from '@material-ui/core/InputLabel';
-import { addProp, deleteProp } from '../actions/components';
+import { addProp, deleteProp } from '../actions/actionCreators';
 import DataTable from './DataTable';
 import { PropInt, PropsInt } from '../utils/Interfaces';
