+++ src/reducers/componentReducer.ts
@@ -19,15 +19,12 @@ import {
   CHANGE_IMAGE_SOURCE,
   DELETE_IMAGE,
   EXPORT_FILES,
-  CREATE_APPLICATION, //unused
   EXPORT_FILES_SUCCESS,
   EXPORT_FILES_ERROR,
-  CREATE_APPLICATION_ERROR, //unused
   HANDLE_CLOSE,
   HANDLE_TRANSFORM,
   OPEN_EXPANSION_PANEL,
   DELETE_ALL_DATA,
-  CHANGE_IMAGE_PATH, //unused
   CHANGE_TUTORIAL,
   ADD_PROP,
   DELETE_PROP,
